LineChatBot/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.1.0
bcrypt==4.0.1
beautifulsoup4==4.8.1
blinker==1.5
certifi==2019.9.11
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
colorama==0.4.6
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
dbus-python==1.2.16
decorator==5.1.1
distlib==0.3.6
distro==1.8.0
distro-info==1.0
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.4
fasteners==0.18
filelock==3.8.0
Flask==1.0.2
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.1
html5lib==1.1
httplib2==0.21.0
idna==2.8
importlib-metadata==5.1.0
iniconfig==1.1.1
itsdangerous==1.1.0
jaraco.classes==3.2.3
jdcal==1.4.1
jeepney==0.8.0
Jinja2==2.10.3
keyring==23.11.0
kiwisolver==1.4.4
launchpadlib==1.10.18
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
line-bot-sdk==1.14.0
lockfile==0.12.2
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.2.4
MarkupSafe==1.1.1
matplotlib==3.6.2
monotonic==1.6
more-itertools==9.0.0
multidict==6.0.2
netifaces==0.11.0
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
openpyxl==3.0.10
packaging==21.3
pandas==0.25.3
paramiko==2.12.0
pexpect==4.8.0
Pillow==9.3.0
pipenv==2022.11.25
pipreqs==0.4.11
platformdirs==2.5.4
pluggy==1.0.0
ptyprocess==0.7.0
py==1.11.0
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
Pygments==2.13.0
PyGObject==3.36.0
pygraphviz==1.5
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyRFC3339==1.1
pytest==7.2.0
python-dateutil==2.8.2
python-debian==0.1.49
python-dotenv==0.10.3
pytz==2022.6
pyxdg==0.28
PyYAML==6.0
reportlab==3.6.12
requests==2.22.0
requests-unixsocket==0.3.0
scipy==1.9.3
SecretStorage==3.3.3
simplejson==3.18.0
six==1.12.0
soupsieve==2.3.2.post1
systemd-python==234
tables==3.7.0
tomli==2.0.1
transitions==0.6.9
urllib3==1.25.6
virtualenv==20.17.0
virtualenv-clone==0.5.7
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.0
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
yarl==1.8.1
zipp==3.11.0
